% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{GCA_clean_text}
\alias{GCA_clean_text}
\title{Clean text}
\usage{
GCA_clean_text(vec)
}
\arguments{
\item{vec}{A character vector}
}
\description{
This function cleans texts in a character vector, erasing links, emojis, mentions and hashtags.
}
