% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{GCA_clean_tweets}
\alias{GCA_clean_tweets}
\title{Clean tweets}
\usage{
GCA_clean_tweets(df, year = ".")
}
\arguments{
\item{df}{A rtweet data.frame (timeline, tweets collected, etc.). No default.}

\item{year}{Year to filter your tweets. Default is ".", extracting all data from your df.}
}
\description{
This function cleans tweets from a rtweet data.frame, erasing links, emojis, mentions and hashtags, which are saved
already in your data.frame. It can also filter tweets from a certain year.
}
