% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna.R
\name{GCA_create_edges_df}
\alias{GCA_create_edges_df}
\title{Create edges data.frame}
\usage{
GCA_create_edges_df(df)
}
\arguments{
\item{df}{A df from GCA_create_term_matrix}
}
\description{
Creates an edges df from your term filtered matrix. Use this method preferably with a small sample.
}
