% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sna.R
\name{GCA_create_term_matrix}
\alias{GCA_create_term_matrix}
\title{Create term matrix}
\usage{
GCA_create_term_matrix(df, term, agent)
}
\arguments{
\item{df}{A matrix generated from GCA_get_corpus}

\item{term}{A character element used to filter your corpus}

\item{agent}{Your agent name, to make it easier to link them}
}
\description{
Filter a matrix to limit itself to a chosen term
}
