% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{GCA_get_corpus}
\alias{GCA_get_corpus}
\title{Get corpus}
\usage{
GCA_get_corpus(vec, stopwords_chos = tm::stopwords("pt"))
}
\arguments{
\item{vec}{Your text to be cleaned and transformed into a corpus}

\item{stopwords_chos}{A character vector with stopwords to be removed}
}
\description{
This function cleans a vector which contains your text to be analyzed, returning a corpus in a dtm format
It's basically a wrapper of different function from the tm package
}
