% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{GCA_get_main_terms}
\alias{GCA_get_main_terms}
\title{Get main terms}
\usage{
GCA_get_main_terms(fir_mx, n = 10)
}
\arguments{
\item{fir_mx}{Your text to be cleaned and transformed into a corpus}

\item{n}{An integer value, defaulted to 10, to choose how much main terms your are looking for}
}
\description{
This function returns the top main terms from a corpus matrix (supports dtm, matrix and data.frame) in a vector.
}
