% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{GCA_get_main_terms_prop}
\alias{GCA_get_main_terms_prop}
\title{Get main terms proportion}
\usage{
GCA_get_main_terms_prop(mx, main_terms)
}
\arguments{
\item{mx}{matrix to be analyzed each main terms value}

\item{main_terms}{a character vector with the main terms to be searched into our matrix}
}
\description{
This function returns the proportions of each main term into the matrix passed into our function.
Returns a named numeric vector
}
