% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{GCA_get_video_corpus}
\alias{GCA_get_video_corpus}
\title{Get Video Corpus}
\usage{
GCA_get_video_corpus(chosen_url, lang = "pt")
}
\arguments{
\item{chosen_url}{Url from a youtube video}

\item{lang}{Language to find transcripts}
}
\description{
This function retrieves the transcript from a video and returns it as a character vector.
}
\details{
Requires: dplyr, youtubecaption
}
